requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.28.0
langfuse>=2.0.0
openai>=1.0.0
huggingface_hub>=0.20.0
qwen-vl-utils>=0.0.10
scenedetect>=0.6
imageio-ffmpeg>=0.5.0
md2pdf
# TransNetV2 (optional, for split_scenes_transnetv2.py)
# transnetv2-pytorch>=1.0.0
